Build a "cdrom" installer and not a "netboot" one
In the live-build code, netinst and netboot have the same meaning,
and it means netboot, an installer that fetches the d-i udebs over
the network. But we don't want that, we want the udeb to be available
in the ISO so that it keeps working even after a kernel update in
kali-rolling.