Fix installer image to cope with default 686-pae kernel that we use on i386
Debian defaults to linux-image-686 for i386 on all levels (simple-cdd,
debian-cd, debian-installer). Kali modified debian-installer to use
linux-image-686-pae in debian-installer (for the benefit of the live
image mainly!). We have to adjust simple-cdd (by way of the
kernel_packages configuration item) and debian-cd (with some live
patching of a copy).