Switch back installer to live one
The netinst installer is causing too many issues for unexperienced
users that are installing without enabling the network. They are ending
up in a console and are stuck there. See https://bugs.kali.org/view.php?id=6053
and the many duplicates.

The live installer is not offering close control on what metapackage
gets installed but at least you are sure to have a usable system at the
end.