grub2: only add speech synth install entry if missing
live-build commit 24aab15d2a133075992ce3a216b1c3476dc7bd4c has added
generation of speech synth entries. this ensures that with the next
release of live-build, the kali one is not added as a duplicate.

we have to take a different approach here than for syslinux. the approach
relies upon an expectation that in the next live-build release the grub
config will include an install.cfg file containing install entries, like
syslinux, as a result of changes awaiting merge in MR #135 ([1]).

[1]: https://salsa.debian.org/live-team/live-build/-/merge_requests/135