grub2: split up grub.cfg
as done in upstream live-build MR #135.

the intention here is to prepare to fix application of the theme for submenus,
as addressed next.

we could accomplish that just by the theme.cfg portion of the split, but why
not do the config.cfg bit also now. this is similar to how syslinux is broken
up, and a part of the live-build MR #135 changes are to merge config/bootloader
files on top of the default set, rather than use them instead, thus
user/derivative configs only need to carry the bootloader files they actually
want to replace. breaking up grub.cfg helps towards this, so eventually kali
can drop carrying files that are identical to upstream ilve-build's.