grub2: fix application of theme to submenus
the theme needs explicitly applying for submenus otherwise they are just
displayed with a default layout with a black background.

with the theme setting commands moved to a separate file, we can now
simply `source` that into each submenu.

this of course assumes that live-theme/theme.txt is suitable for submenus.