Switch DISKTYPE to BD for Bluray Disk
The DVD size limit has been reached and the build currently fails due to
this. We don't really want to burn DVD as ISO images are written on USB
keys and the exact size doesn't really matter for use in virtual
machines. So let's instruct debian-cd to build a bluray image so that
we no longer are annoyed with the size limit.