Add kali user to additional groups one by one
Up to now, it was fine to use a command such as:

    usermod -a -G group1,group2,... kali

However a limitation is that all the groups that are given to the option
-G must exist. If that's not the case, usermod fails (return code: 6)
without doing anything, and the user is not added to any group.

So with this commit, we prepare the code to support optional groups,
that might or might not exist. If ever a group does not exist, it's
skipped silently.