Disable the encfs error message
Ths first line ('... boolean true') is needed to trick debian-cd into
believing that the type of the question is a boolean.

Otherwise debian-cd will think that the question doesn't exist, probably
because it doesn't know the type, probably because there's no type to
start with (this is an error message, not a question).

FTR, without the first line, it fails with:

  2022-08-04 10:05:37,846 DEBUG Checking configuration...
  error: Cannot find a question for encfs/security-information
  2022-08-04 10:05:38,016 ERROR preseed file invalid:
    <<LBDIR>>/simple-cdd/profiles/kali.preseed