Enable Kali Purple installer theme
Other things that I tried, but that didn't work, below:

I tried adding the preseed to simple-cdd/profiles/kali-purple.preseed,
however that didn't work, I guess that this preseed file is loaded only
later during installation, and that at this point the GUI was already
loaded and it's too late to set the theme.

I also tried to do this in simple-cdd.conf:

  # Theming
  if echo " $profiles " | grep -q " kali-purple "; then
    export KERNEL_PARAMS="${KERNEL_PARAMS} debian-installer/theme=Clearlooks-Purple "
  fi

But it doesn't work, seemingly 'profiles' is not set when simple-cdd
loads the conf, and even though we build with the command:

    build-simple-cdd [...] --profiles "$profiles" [...]

Given the fact that uppercase variables are exported, and what we
actually want to change is KERNEL_PARAMS, it seems that the most
straightforward is just to preset it from build.sh.