Kali purple: install kali-system-gui (instead of core)
The dep chain is kali-system-gui > kali-system-cli > kali-system-core.

kali-system-cli contains wget and curl, which we probably want to have,
and will fix https://gitlab.com/kalilinux/kali-purple/documentation/-/issues/9

I'm not sure about kali-system-gui, but I have the impression that it
was the intention to install it in Kali Purple, but at some point it was
forgotten and we install kali-system-core instead.