Revert "Make sure to include kali-linux-firmare"
This reverts commit 7bb52e4991b135903e30d74439635c94f93ed592.

I wanted to make sure firmware are included in the installer image, but
I missed it, I modified a config file that is only for live images, it
seems...

Also, note that the previous commit message

  Kali purple: install kali-system-gui (instead of core)

Should have been:

  Kali purple: include kali-system-gui (instead of core) in the installer

Because this commit is really about having the package available in the
installer, but it doesn't "install it" on the system. Sorry for the
confusion.