Make sure to include kali-linux-firmware in the installer images
Usually it's pulled in via dependency (as it's a dep of
kali-linux-headless IIRC itself a dep of kali-linux-default). But that's
not the case for Kali Purple, as it doesn't depend on
kali-linux-default.

Furthermore, kali-linux-firmware is now installed in all cases by
tasksel, so we must make sure it's there, instead of relying on
metapackage's dependencies.

Hopefully it will fix:
- https://gitlab.com/kalilinux/kali-purple/documentation/-/issues/8
- https://gitlab.com/kalilinux/kali-purple/documentation/-/issues/11